finalized edit, validate, and delete

+++ Rubio_Esau_Project3/js/main.js
@@ -1,7 +1,7 @@
-// Project 2 : WOrkganizer
+// Project 3 : WOrkganizer
 // Author: Esau Rubio
 // VFW 0113
-// 1/15/13
+// 1/23/13
 
 // DOM Loaded
 window.addEventListener( "DOMContentLoaded" , function() {
@@ -13,7 +13,7 @@ window.addEventListener( "DOMContentLoaded" , function() {
     
     // Save Data
     function getPriority() {
-        var radio = document.forms[0].priorit;
+        var radio = document.forms[0].priority;
         console.log(radio);
         console.log(radio.checked);
         for (var i=0 , rl=radio.length; i<rl ; i++) {
@@ -22,12 +22,16 @@ window.addEventListener( "DOMContentLoaded" , function() {
             };
         };
     };
-    function storeData() {
+    function storeData(key) {
+        if (!key) {
+            var keyGen = Math.floor(Math.random()*1000000);
+        } else {
+            var keyGen = key;
+        };
         getPriority ();
-        var keyGen = Math.floor(Math.random()*1000000);
         var	userInput = {};
             userInput.location = ["Location:" , ids("location").value];
-            userInput.worktype = ["Work Type:" , ids("typeOfWork").value];
+            userInput.worktype = ["Work Type:" , ids("worktype").value];
             userInput.priority = ["Priority:" , priority ];
             userInput.people   = ["Workers Sent:" , ids("people").value];
             userInput.finishby = ["Finish By:" , ids("finishby").value];
@@ -40,30 +44,141 @@ window.addEventListener( "DOMContentLoaded" , function() {
     function getData() {
         switchControl("on")
         var makeDiv = document.createElement('div');
-            makeList = document.createElement('ul');
         makeDiv.setAttribute('id', 'list');
-        makeDiv.appendChild(makeList);
         document.body.appendChild(makeDiv);
         ids('list').style.display = 'block';
         for(var i=0, l=localStorage.length; i<l; i++) {
-            var makeLi = document.createElement('li');
+            var makeList = document.createElement('ul');
+                makeLink = document.createElement('li'); 
                 key = localStorage.key(i);
                 value = localStorage.getItem(key);
                 parsed = JSON.parse(value);
-                subUl = document.createElement('ul');
                 makeBreak = document.createElement('br');
-            makeList.appendChild(makeLi)
-            makeLi.appendChild(subUl)
             for(var n in parsed) {
-                var subLi = document.createElement('li');
+                var moreLi = document.createElement('li');
                     text = parsed[n][0] + " " + parsed[n][1];
-                subUl.appendChild(subLi);
-                subLi.innerHTML = text;
+                makeList.appendChild(moreLi);
+                moreLi.innerHTML = text;
             };
-        makeLi.appendChild(makeBreak);
+            makeDiv.appendChild(makeList);
+            makeList.appendChild(makeLink);
+            makeList.appendChild(makeBreak);
+            editDeleteLinks(localStorage.key(i) , makeLink);
         };
     };
     
+    // Create Edit and Delete Links
+    function editDeleteLinks(key , makeLink) {
+        var edit = document.createElement('a');
+            edit.href = '#';
+            edit.key = key;
+            editText = "Edit Job";
+        edit.addEventListener( 'click' , editor);
+        edit.innerHTML = editText;
+        makeLink.appendChild(edit);
+        
+        var del = document.createElement('a');
+            del.href = '#';
+            del.key = key;
+            delText = "Delete Job";
+        del.addEventListener( 'click' , deletor);
+        del.innerHTML = delText;
+        makeLink.appendChild(del);
+    };
+    
+    // Edit a choosen Job
+    function editor() {
+        var getInput = localStorage.getItem(this.key);
+            userInput = JSON.parse(getInput);
+        switchControl("off")
+        ids('location').value = userInput.location[1];
+        ids('worktype').value = userInput.worktype[1];
+        ids('people').value = userInput.people[1];
+        ids('finishby').value = userInput.finishby[1];
+        ids('notes').value = userInput.notes[1];
+        var rad = document.forms[0].priority;
+        for (var i=0 , len=rad.length; i<len; i++) {
+            if (rad[i].value == "High" && userInput.priority[1] == "High") {
+                rad[i].setAttribute('checked' , 'checked')
+            } else if (rad[i].value == "Medium" && userInput.priority[1] == "Medium") {
+                rad[i].setAttribute('checked' , 'checked')
+            } else if (rad[i].value == "Low" && userInput.priority[1] == "Low") {
+                rad[i].setAttribute('checked' , 'checked')
+            };
+        };
+        submit.removeEventListener('click' , storeData);
+        var editSubmit = ids('submit');
+        editSubmit.addEventListener('click' , validate);
+        editSubmit.key = this.key;
+    };
+    
+    // Delete a choosen Job
+    function deletor() {
+        var ask = confirm('Are you Sure?');
+        if (ask) {
+            localStorage.removeItem(this.key)
+            window.location.reload();
+            alert('Job Deleted!')
+        } else {
+            alert('Job NOT Deleted')
+        };
+    };
+    
+    // Validate the user input
+    function validate(e) {
+        var errorArray = [];
+            getLocation = ids('location');
+            getWorktype = ids('worktype');
+            getFinishby = ids('finishby');
+            getNotes = ids('notes');
+            
+        // Reset
+        errorMessage.innerHTML = '';
+        getWorktype.style.border = '1px solid black';
+        getLocation.style.border = '1px solid black';
+        getFinishby.style.border = '1px solid black';
+        getNotes.style.border = '1px solid black';
+        
+        // Work Type Validation
+        if (getWorktype.value==='-Maintenance-') {
+            var worktypeError = "Please choose a valid Work Type";
+            getWorktype.style.border = '1px solid red'
+            errorArray.push(worktypeError)
+        };
+        // Location Validation
+        if (getLocation.value === '') {
+            var locationError = "Please choose a valid Location";
+            getLocation.style.border = '1px solid red'
+            errorArray.push(locationError)
+        };
+        // Date validation
+        var checkDate = /^\d{4}[\/\-](0?[1-9]|1[012])[\/\-](0?[1-9]|[12][0-9]|3[01])$/;
+        if (!(checkDate.exec(getFinishby.value))) {
+            var finishbyError = "Please choose a valid Finish Date";
+            getFinishby.style.border = '1px solid red'
+            errorArray.push(finishbyError)
+        };
+        // Notes Validation
+        if (getNotes.value === '') {
+            var notesError = "Please enter Notes";
+            getNotes.style.border = '1px solid red'
+            errorArray.push(notesError)
+        };
+        // Display Error Message
+        if(errorArray.length>0) {
+            for(var i=0,len=errorArray.length; i<len; i++) {
+                var info = document.createElement('li')
+                info.innerHTML = errorArray[i]
+                errorMessage.appendChild(info)
+            };
+            e.preventDefault();
+            return false;
+        } else {
+            storeData(this.key);
+        };
+    };
+    
+    // Switch between Add Job and Display Jobs
     function switchControl(n) {
         switch(n) {
             case "on":
@@ -88,19 +203,21 @@ window.addEventListener( "DOMContentLoaded" , function() {
         if (localStorage.length===0) {
             alert("No Jobs to clear!");
         } else {
-            localStorage.clear();
-            confirm("Are you sure you want to clear all Jobs??");
-            alert("Jobs Cleared!");
-            return false;
+            var areYouSure = confirm("Are you sure you want to clear all Jobs??");
+            if(areYouSure ==  true) {
+                localStorage.clear();
+                alert("Jobs Cleared!");
+                return false;
+            };
         };
     };
 
     // Dynamic Options
     function addCat() {
         var selectForm = document.getElementsByTagName("form"),
-            selectLi = ids("worktype");
+            selectLi = ids("typeOfWork");
         selection = document.createElement("select");
-        selection.setAttribute( "id" , "typeOfWork" );
+        selection.setAttribute( "id" , "worktype" );
         for(var i=0, n=maintenanceTypes.length ; i<n ; i++ ) {
             var makeOption = document.createElement("option"),
                 text = maintenanceTypes[i];
@@ -114,6 +231,7 @@ window.addEventListener( "DOMContentLoaded" , function() {
     // Default Values
     var maintenanceTypes = [ "-Maintenance-", "Maintenace: House Keeping" , "Maintenace: Painting", "Maintenace: Electric" , "Maintenace: Plumbing"],
         priority;
+        errorMessage = ids('error')
 		
     addCat ()
     //Clear and Display Data
@@ -122,5 +240,5 @@ window.addEventListener( "DOMContentLoaded" , function() {
         submit = ids("submit");
     display.addEventListener( "click" , getData );
     clear.addEventListener( "click" , clearData );
-    submit.addEventListener( "click" , storeData );
+    submit.addEventListener( "click" , validate );
 });
